\name{across.reps-methods}
\docType{methods}
\alias{across.reps-methods}
\alias{across.reps}
\alias{across.reps,classification-method}
\title{Summarises classification values across bugs or jags replications}
\description{
Summarises classification values across bugs or jags replications
}
\section{Methods}{
\describe{

\item{\code{signature(object = "classification")}}{

}
}}
