\name{beta.list}
\alias{beta.list}
\title{
Extract Beta Values from Bugs Sims File
}
\description{
Extracts beta values from bugs sims file.
}
\usage{
beta.list(v)
}
\arguments{
  \item{v}{
  Bugs sims file
}
}
\details{
Draws heavily on Curtis, S.M.(2010). 
}
\value{
Returns list of beta parameters.
}
\references{
Curtis, S.M.(2010) BUGS Code for Item Response Theory, \emph{Journal of Statistical Software}, Code Snippets, \bold{36(1)},1--34
}