\name{cv_partition}
\alias{cv_partition}
\title{Partitions data for cross-validation...}
\usage{cv_partition(y, num_folds=5)}
\description{Partitions data for cross-validation}
\value{list the indices of observations to hold out by fold..}
\arguments{\item{y}{a vector of the labels of the training data}
\item{k}{leave out how many observations per fold in cross-validation?}}
