\name{error_632plus}
\alias{error_632plus}
\title{Classification error rate estimation via the 632+ estimator}
\usage{
  error_632plus(d, list_classifiers, num_folds = 5, ...)
}
\arguments{
  \item{d}{TODO}

  \item{list_classifiers}{TODO}

  \item{num_folds}{TODO}
}
\value{
  error_results TODO
}
\description{
  TODO
}

