\name{scores-class}
\Rdversion{1.1}
\docType{class}
\alias{scores-class}
\alias{scores,missing-method}

\title{Class \code{"scores"}}
\description{
S4 object for score distributions.
}
\section{Slots}{
  \describe{
    \item{\code{item}:}{Item scores}
    \item{\code{expected}:}{Expected scores}
    \item{\code{conditional}:}{Conditional Summed scores}
    \item{\code{summed}:}{Summed scores}
    \item{\code{persons}:}{Number of persons}
    \item{\code{items}:}{Number of items}
    \item{\code{sims}:}{Number of simulations}
    \item{\code{max}:}{Maximum test score}
    \item{\code{cats}:}{Item categories}
    \item{\code{model}:}{Bugs model}
    \item{\code{estimation}:}{Bugs estimation software}
  }
}
