\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,classification-method}
\alias{summary}
\title{Summary Statistics for S4 Class Classification}
\description{Summary Statistics for S4 Class Classification}
\section{Methods}{
\describe{
\item{\code{signature(object = "classification")}}{
Produces summary statistics from the S4 classification object.
}
}}