\name{error_632}
\alias{error_632}
\title{Classification error rate estimation via the 632 estimator...}
\usage{error_632(d, list_classifiers, num_folds=5, ...)}
\description{Classification error rate estimation via the 632 estimator}
\details{TODO}
\value{error_results TODO}
\arguments{\item{d}{TODO}
\item{list_classifiers}{TODO}
\item{num_folds}{TODO}}
