\name{error_apparent}
\alias{error_apparent}
\title{Classification error rate estimation via the apparent error rate estimator.}
\usage{error_apparent(d, list_classifiers, ...)}
\description{Classification error rate estimation via the apparent error rate estimator.}
\details{TODO}
\value{error_results TODO}
\arguments{\item{d}{TODO}
\item{list_classifiers}{TODO}}
