\name{error_boot}
\alias{error_boot}
\title{Classification error rate estimation via bootstrapping}
\usage{
  error_boot(d, list_classifiers, num_reps = 20, ...)
}
\arguments{
  \item{d}{TODO}

  \item{list_classifiers}{TODO}

  \item{num_reps}{TODO}
}
\value{
  error_results TODO
}
\description{
  TODO
}

