\name{error_cv}
\alias{error_cv}
\title{Classification error rate estimation via cross-validation}
\usage{
  error_cv(d, list_classifiers, num_folds = 5, vs_method =
  NULL, ...)
}
\arguments{
  \item{d}{TODO}

  \item{list_classifiers}{TODO}

  \item{num_folds}{TODO}
}
\value{
  error_results TODO
}
\description{
  TODO
}

