\name{get_cl_attrib}
\alias{get_cl_attrib}
\title{Retrieves the functions for the specified classifier.}
\usage{
  get_cl_attrib(cl)
}
\arguments{
  \item{cl}{TODO}
}
\value{
  list TODO
}
\description{
  TODO
}
\details{
  Example: cl = list( method = "RDA", classifier = "rda",
  model_select = "rda_model_select", predict =
  "predict_rda" )
}

